SpecDeploymentConsumer: catch more upload errors (& retry more on deploy failure)
Seeing a lot of networking issues, this is trying to make
upload a bit more tolerant to network issues.

There are two cases where we can get IOExceptions:
* from the actual apache httpclient org.apache.http.client CloseableHttpClient
* as a result of parsing the Artifactory response (JSON parsing) in
  private ArtifactoryUploadResponse uploadFile(DeployDetails details, String uploadUrl)

The 1st case is handled by existing code, but the 2nd case
isn't so far. This commit adds handling to the JSON parsing.

Signed-off-by: André Draszik <andre.draszik@jci.com>