Refactor build info extraction to be safer
This refactors build info extraction so that it doesn't need to use
internal Gradle classes and occurs safely in parallel.  Build info is
now calculated for each module using a task in the associated project
and shared with the root project via a variant artifact.  The root
project then aggregates the module info and publishes it with the
overall build info..