Switch to case insensitive comparison of HTTP headers
A bug was introduced in 67cefb2 where http headers no longer was
compared case insensitively. This causes problems for Artifactory
instances that run behind proxys that normalizes the http headers.

This change fixes the issue by mimicking the implementation in
org.apache.http.message.HeaderGroup by using String#equalsIgnoreCase().
Which is the old and correct behaviour.